requirements.txt
numpy~=1.26.4
pillow~=10.3.0
openmeteo-requests~=1.2.0
requests-cache~=1.2.0
retry-requests~=2.0.0
pandas~=2.2.2

pygame~=2.5.2